# src/M'baara-openapi-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "M'baara API",
    "description": "API documentation for the **M'baara** application.\n\n**Authentication:** Most endpoints require the `api_key` header. If using Swagger UI, click the **Authorize** button at the top of the page and enter your API key before trying any endpoint.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000/api",
      "description": "Local API server"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/entities/Lesson": {
      "get": {
        "tags": [
          "Entities - Lesson"
        ],
        "summary": "List Lesson records",
        "operationId": "list_Lesson",
        "description": "Retrieve a list of Lesson records with optional filtering, sorting, and pagination.",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "JSON query filter, e.g. {\"status\":\"active\"}"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 100
            },
            "description": "Maximum number of records to return"
          },
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            },
            "description": "Number of records to skip (pagination)"
          },
          {
            "name": "sort_by",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Field name to sort by. Prefix with '-' for descending order, e.g. -created_date"
          }
        ],
        "responses": {
          "200": {
            "description": "List of records",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Lesson"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Entities - Lesson"
        ],
        "summary": "Create a Lesson record",
        "operationId": "create_Lesson",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Lesson"
              },
              "example": {
                "language_code": "Example language_code",
                "title": "Example title",
                "order": 0
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lesson"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Entities - Lesson"
        ],
        "summary": "Delete multiple Lesson records",
        "operationId": "delete_many_Lesson",
        "description": "**Warning:** Passing an empty `{}` query will delete ALL records in this entity.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Query filter to match records to delete"
              },
              "example": {
                "language_code": "Example language_code"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Deletion result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "deleted": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/entities/Lesson/{Lesson_id}": {
      "get": {
        "tags": [
          "Entities - Lesson"
        ],
        "summary": "Get a Lesson record by ID",
        "operationId": "get_Lesson",
        "parameters": [
          {
            "name": "Lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Record ID"
          }
        ],
        "responses": {
          "200": {
            "description": "The record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lesson"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Entities - Lesson"
        ],
        "summary": "Update a Lesson record",
        "operationId": "update_Lesson",
        "parameters": [
          {
            "name": "Lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Record ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Lesson"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lesson"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Entities - Lesson"
        ],
        "summary": "Delete a Lesson record",
        "operationId": "delete_Lesson",
        "description": "Soft-deletes the record. Use the restore endpoint to undo.",
        "parameters": [
          {
            "name": "Lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Record ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Deletion result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/entities/User": {
      "get": {
        "tags": [
          "Entities - User"
        ],
        "summary": "List User records",
        "operationId": "list_User",
        "description": "Retrieve a list of User records with optional filtering, sorting, and pagination.",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "JSON query filter, e.g. {\"status\":\"active\"}"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 100
            },
            "description": "Maximum number of records to return"
          },
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            },
            "description": "Number of records to skip (pagination)"
          },
          {
            "name": "sort_by",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Field name to sort by. Prefix with '-' for descending order, e.g. -created_date"
          }
        ],
        "responses": {
          "200": {
            "description": "List of records",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Entities - User"
        ],
        "summary": "Create a User record",
        "operationId": "create_User",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              },
              "example": {
                "email": "Example email",
                "full_name": "Example full_name",
                "role": "admin"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
    "/entities/User/{User_id}": {
      "get": {
        "tags": [
          "Entities - User"
        ],
        "summary": "Get a User record by ID",
        "operationId": "get_User",
        "parameters": [
          {
            "name": "User_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Record ID"
          }
        ],
        "responses": {
          "200": {
            "description": "The record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Entities - User"
        ],
        "summary": "Update a User record",
        "operationId": "update_User",
        "parameters": [
          {
            "name": "User_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Record ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Entities - User"
        ],
        "summary": "Delete a User record",
        "operationId": "delete_User",
        "description": "Soft-deletes the record. Use the restore endpoint to undo.",
        "parameters": [
          {
            "name": "User_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Record ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Deletion result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "api_key",
        "description": "Your API key. Obtain it from your user profile or via `User.me()` in the app SDK."
      }
    },
    "schemas": {
...
      "Lesson": {
        "type": "object",
        "properties": {
          "language_code": {
            "type": "string"
          },
          "lesson_number": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "title_fr": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "level": {
            "type": "string",
            "enum": [
              "A1",
              "A2",
              "B1",
              "B2"
            ],
            "default": "A1"
          },
          "order": {
            "type": "number"
          },
          "type": {
            "type": "string",
            "enum": [
              "vocabulary",
              "phrases",
              "letters",
              "sounds"
            ],
            "default": "vocabulary"
          },
          "is_published": {
            "type": "boolean",
            "default": true
          },
          "id": {
            "type": "string",
            "description": "Unique record identifier"
          },
          "created_date": {
            "type": "string",
            "format": "date-time",
            "description": "Record creation timestamp"
          },
          "updated_date": {
            "type": "string",
            "format": "date-time",
            "description": "Record last update timestamp"
          },
          "created_by_id": {
            "type": "string",
            "description": "ID of the user who created the record"
          }
        },
        "required": [
          "language_code",
          "title",
          "order"
        ],
        "description": "Lesson entity"
      },
      "User": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "The email of the user"
          },
          "full_name": {
            "type": "string",
            "description": "The full name of the user"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "user"
            ],
            "description": "The role of the user in the app"
          },
          "id": {
            "type": "string",
            "description": "Unique record identifier"
          },
          "created_date": {
            "type": "string",
            "format": "date-time",
            "description": "Record creation timestamp"
          },
          "updated_date": {
            "type": "string",
            "format": "date-time",
            "description": "Record last update timestamp"
          },
          "created_by_id": {
            "type": "string",
            "description": "ID of the user who created the record"
          }
        },
        "required": [
          "email",
          "role",
          "full_name"
        ],
        "description": "User entity"
      }
    }
  },
...
}